Adjust the timing of configuring ElasticJob in Spring Boot startup process (#1703)
