+++ elasticjob-lite/elasticjob-lite-spring/elasticjob-lite-spring-boot-starter/src/main/java/org/apache/shardingsphere/elasticjob/lite/spring/boot/job/ScheduleJobBootstrapStartupRunner.java
@@ -29,7 +29,7 @@ import org.springframework.context.ApplicationContextAware;
  */
 @Setter
 @Slf4j
-public class ElasticJobStartupRunner implements CommandLineRunner, ApplicationContextAware {
+public class ScheduleJobBootstrapStartupRunner implements CommandLineRunner, ApplicationContextAware {
     
     private ApplicationContext applicationContext;
